pyproject.toml
[tool.poetry]
name = "deluminate"
version = "0.1.0"
description = ""
authors = ["RyanTWalsh <rwalsh@x-chemrx.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
numpy = "^1.26.2"
pyarrow = "^14.0.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
rdkit = "^2023.9.3"
metaflow = "^2.10.8"
scikit-learn = "^1.3.2"
python-ternary = "^1.0.8"
ugtm = "^2.1.0.post2"
altair = "^5.2.0"
jupyter = "^1.0.0"
SciPy = "^1.11.4"
MiniSom = "^2.3.1"
mhfp = "^1.9.6"
mxfp = "^1.1.2"
networkx = "^3.2.1"
tqdm = "^4.66.1"
faerun = "^0.4.7"
map4 = {git = "https://github.com/reymond-group/map4", rev = "v1.0"}
openpyxl = "^3.1.2"
umap-learn = "^0.5.5"
XlsxWriter = "^3.1.9"
annoy = "^1.17.3"
tmap-viz = "^1.0.16"
PyTDC = "^0.4.1"
oracledb = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
